+++ main.py
@@ -1,7 +1,12 @@
 import pygame, random, time, threading
-from animations import PlayerAnimations
+from objects.animations import PlayerAnimations
 from perlin_noise import PerlinNoise
 import pathlib, pickle, cv2
+
+
+from objects.player import Player
+from objects.utils import FPScounter, log
+from objects.animations import PlayerAnimations
 
 StartTime=time.time()
 MAP_SIZE = 6400
@@ -17,10 +22,20 @@ COLORS = {
     3: (85, 86, 87), # rocks
     4: (34, 107, 34), # Plains higher
     5: (139, 137, 137), # Mountains
-    6: (255, 250, 250) # Snowy Mountains,
-
+    6: (255, 250, 250), # Snowy Mountains,
+    7: "tree"
 }
 
+STRUCTURES = {
+    "tree": [
+    pygame.image.load("./resources/assets/tree0.png"),
+    pygame.image.load("./resources/assets/tree1.png"),
+    pygame.image.load("./resources/assets/tree2.png"),
+    pygame.image.load("./resources/assets/tree3.png"),
+    pygame.image.load("./resources/assets/tree4.png"),
+    pygame.image.load("./resources/assets/tree5.png"),
+    ]
+}
 def log(*args, **kw): # Debug
     print(*args, **kw)
 
@@ -49,14 +64,14 @@ class Tilemap:
                 elif height < 7:
                     self.map[i][j] = 1
                 elif height < 40:
-                    self.map[i][j] = self.random(2, 3, 500)
+                    self.map[i][j] = 3 if self.random(2, 3, 500) == 3 else self.random(2, 7, 500) if 20 < height < 40 else 2
                 elif height < 60:
-                    self.map[i][j] = self.random(4, 3, 500)
+                    self.map[i][j] = 3 if self.random(4, 3, 500) == 3 else self.random(4, 7, 500)
                 elif height < 80:
                     self.map[i][j] = 5
                 else:
                     self.map[i][j] = 6
-                
+
         log("Terrain Generated")
 
 
@@ -65,19 +80,11 @@ class Tilemap:
         for x in range(round(player.x - 44), round(player.x + 44)):
             for y in range(round(player.y - 26), round(player.y + 26)): # CLIPPING VALUES. TO CHANGE
                 tile = self.map[x][y]
-                pygame.draw.rect(surface, COLORS[tile] if x != round(player.x) or y != round(player.y) else (255, 0, 0), pygame.Rect(x * 16 - round(player.x * 16) + SCREEN_WIDTH // 2, y * 16 - round(player.y * 16) + SCREEN_HEIGHT // 2, 32, 32))
-
-class FPScounter:
-    def __init__(self, clock, screen, player):
-        self.clock = clock
-        self.screen = screen
-        self.player = player
-
-    def display(self):
-        font = pygame.font.SysFont(None, 30)
-        img = font.render(f"{round(self.clock.get_fps())}, X: {round(self.player.x, 2)}, Y: {round(self.player.y, 2)}", True, (0, 0, 0))
-        self.screen.blit(img, (2, 2))
-        
+                if tile < 7:
+                    pygame.draw.rect(surface, COLORS[tile] if (x != round(player.x) or y != round(player.y)) else (255, 0, 0), pygame.Rect(x * 16 - round(player.x * 16) + SCREEN_WIDTH // 2, y * 16 - round(player.y * 16) + SCREEN_HEIGHT // 2, 32, 32))
+                else:
+                    surface.blit(pygame.transform.scale(STRUCTURES[COLORS[tile]][0], (128, 128)), (x * 16 - 80 - round(player.x * 16)  + (SCREEN_WIDTH // 2 ), y * 16 - 100 - round(player.y * 16)  + (SCREEN_HEIGHT // 2))) # joueur toujours au millieu de l'écran, c'est le bg qui bouge
+  
 class Entity:
     def __init__(self, coords, screen):
         self.x, self.y = coords
@@ -87,141 +94,6 @@ class Entity:
         self.speed = 16
         self.screen = screen
         
-       
-def collide(c1, c2, velocity):
-    # self: the dynamic collider, which moves
-    # collider: the static collider, which stays put
-    
-    c1x1, c1y1, c1x2, c1y2 = c1
-    c2x1, c2y1, c2x2, c2y2 = c2
-    
-    no_collision = 1, None
-
-    # find entry & exit times for each axis
-
-    vx, vy = velocity
-
-    time = lambda x, y: x / y if y else float('-' * (x > 0) + "inf")
-
-    x_entry = time(c2x1 - c1x2 if vx > 0 else c2x2 - c1x1, vx)
-    x_exit  = time(c2x2 - c1x1 if vx > 0 else c2x1 - c1x2, vx)
-
-    y_entry = time(c2y1 - c1y2 if vy > 0 else c2y2 - c1y1, vy)
-    y_exit  = time(c2y2 - c1y1 if vy > 0 else c2y1 - c1y2, vy)
-
-    # make sure we actually got a collision
-
-    if x_entry < 0 and y_entry < 0:
-        return no_collision
-
-    if x_entry > 1 or y_entry > 1:
-        return no_collision
-    
-    # on which axis did we collide first?
-
-    entry = max(x_entry, y_entry)
-    exit_ = min(x_exit,  y_exit)
-
-    if entry > exit_:
-        return no_collision
-    
-    # find normal of surface we collided with
-
-    nx = (0, -1 if vx > 0 else 1)[entry == x_entry]
-    ny = (0, -1 if vy > 0 else 1)[entry == y_entry]
-
-
-    return entry, (nx, ny)     
-        
-
-
-
-class Player(pygame.sprite.Sprite):
-    def __init__(self, coords, screen, tilemap):
-        super().__init__()
-        self.tilemap = tilemap
-        self.image = pygame.image.load("./resources/animations/player/idle/idle00.png")
-        self.rect = self.image.get_rect()
-        self.x, self.y = coords
-        self.velocity = [0, 0]
-        self.accel = [0, 0]
-        self.friction = FRICTION
-        self.speed = 16
-        self.screen = screen
-        self.rigidBody = pygame.Rect(self.x, self.y, 16, 16)
-        self.isAttacking = [False, 0, "r"]
-        self.size = 128* 1.25
-        
-    def render(self, skin=pygame.image.load("./resources/animations/player/idle/idle00.png")):
-        self.image = pygame.transform.scale(skin, (self.size, self.size))
-        self.screen.blit(self.image, (SCREEN_WIDTH//2 - (self.size//2), SCREEN_HEIGHT//2 - (self.size//2))) # joueur toujours au millieu de l'écran, c'est le bg qui bouge
-   
-    def move(self, delta_time):
-        self.velocity = [v + a * f * delta_time for v, a, f in zip(self.velocity, self.accel, self.friction)]
-      
-        
-        for _ in range(2):
-            adjusted_velocity = [v * delta_time for v in self.velocity]
-            vx, vy = adjusted_velocity
-
-            # find all the blocks we could potentially be colliding with
-            # this step is known as "broad-phasing"
-
-            step_x = 1 if vx > 0 else -1
-            step_y = 1 if vy > 0 else -1
-
-            steps_xz = 1
-            steps_y  = 1
-
-            x, y = int(self.x), int(self.y)
-            cx, cy = [int(x + v) for x, v in zip((self.x, self.y), adjusted_velocity)]
-
-            potential_collisions = []
-
-            for i in range(x - step_x * (steps_xz + 1), cx + step_x * (steps_xz + 2), step_x):
-                for j in range(y - step_y * (steps_y + 2), cy + step_y * (steps_y + 3), step_y):
-                    pos = (i, j)
-                    
-                    if i < 0 and j < 0:
-                        continue
-                    tile = self.tilemap.map[i][j]
-
-                    if tile < 5:
-                        continue
-                    
-                    entry_time, normal = collide((self.x, self.y, self.x + 2, self.y + 2), (i, j, i+1, j+1), adjusted_velocity)
-
-                    if normal is None:
-                        continue
-
-                    potential_collisions.append((entry_time, normal))
-
-            # get first collision
-
-            if not potential_collisions:
-                break
-
-            entry_time, normal = min(potential_collisions, key = lambda x: x[0])
-            entry_time -= 0.001
-
-            if normal[0]:
-                self.velocity[0] = 0
-                self.x += vx * entry_time
-            
-            if normal[1]:
-                self.velocity[1] = 0
-                self.y += vy * entry_time
-
-                        
-              
-        self.x += self.velocity[0] * delta_time * self.speed
-        self.y += self.velocity[1] * delta_time * self.speed
-
-           
-
-        
-        self.velocity = [v - min(v * f * delta_time, v, key = abs) for v, f in zip(self.velocity, self.friction)]
-
 
 class EventHandler:
     def __init__(self, game):
@@ -298,8 +170,8 @@ class Game:
         else:
             self.load()
                         
-        self.screen = pygame.display.set_mode((SCREEN_WIDTH, SCREEN_HEIGHT), pygame.FULLSCREEN)
-        self.player = Player(self.level.player_coords, self.screen, self.tilemap)
+        self.screen = pygame.display.set_mode((SCREEN_WIDTH, SCREEN_HEIGHT), 0)
+        self.player = Player(self.level.player_coords, self.screen, self.tilemap, FRICTION, SCREEN_WIDTH, SCREEN_HEIGHT)
         self.fps_counter = FPScounter(self.clock, self.screen, self.player)
         self.playerAnimations = PlayerAnimations(self.player)
         self.loading = False
@@ -321,7 +193,7 @@ class Game:
             log("Level saved")
 
     def loadingText(self):
-        self.screen = pygame.display.set_mode((SCREEN_WIDTH, SCREEN_HEIGHT), pygame.FULLSCREEN)
+        self.screen = pygame.display.set_mode((SCREEN_WIDTH, SCREEN_HEIGHT), 0)
         font = pygame.font.Font(None, 32)
         text = font.render('Chargement...', True, (255, 255, 255))
         textRect = text.get_rect()
@@ -334,7 +206,7 @@ class Game:
         video = cv2.VideoCapture("./resources/video/Annim_nuages_sans_narration.mp4")
         success, video_image = video.read()
         fps = video.get(cv2.CAP_PROP_FPS)
-        self.screen = pygame.display.set_mode(video_image.shape[1::-1], pygame.FULLSCREEN)
+        self.screen = pygame.display.set_mode(video_image.shape[1::-1], 0)
         
         soundObj = pygame.mixer.Sound('./resources/video/piste_audio.mp3')
         soundObj.play()
