+++ objects/utils.py
@@ -1,0 +1,16 @@
+import pygame
+
+def log(*args, **kw): # Debug
+    print(*args, **kw)
+
+class FPScounter:
+    def __init__(self, clock, screen, player):
+        self.clock = clock
+        self.screen = screen
+        self.player = player
+
+    def display(self):
+        font = pygame.font.SysFont(None, 30)
+        img = font.render(f"{round(self.clock.get_fps())}, X: {round(self.player.x, 2)}, Y: {round(self.player.y, 2)}", True, (0, 0, 0))
+        self.screen.blit(img, (2, 2))
+
